requirements.txt
TA-Lib
flask
panda
yfinance